worker/requirements.txt
python-dotenv
azure-identity
azure-storage-blob
azure-servicebus
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
azure-core-tracing-opentelemetry
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-openai-v2
openai
azure-cosmos
aiohttp
markitdown
pydantic